Modularize registration form
